had to completely rewrite both context, but they working now for some reason...
