+++ reactsignalr/src/index.js
@@ -3,19 +3,19 @@ import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 import './index.css';
 import App from './App';
-import { AuthProvider } from './contexts/auth.js';
-import { HubProvider } from './contexts/hub.js';
+import { UserProvider } from './contexts/userContext';
+import { HubProvider } from './contexts/hubContext';
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
   <React.StrictMode>
-    <AuthProvider>
-      <HubProvider>
-      <BrowserRouter>
-    <App />
+    <BrowserRouter>
+      <UserProvider>
+          <HubProvider>
+            <App />
+          </HubProvider>
+      </UserProvider>
     </BrowserRouter>
-    </HubProvider>
-    </AuthProvider>
   </React.StrictMode>,
   document.getElementById('root')
 );
